oswatcher: move asn1 into utils
